refactor(more 3.0 cleanup): cleanup more inbox 3.0 release
BREAKING CHANGE: markRead, archive, unread, etc...  are updated and TextBlock -> TextElement
