+++ packages/react-inbox/src/components/index.ts
@@ -1,4 +1,4 @@
 export { CourierProvider } from "@trycourier/react-provider";
 export { default as Inbox } from "./Inbox";
 export { default as Header } from "./Messages2.0/Header";
-export { TextBlock, ActionBlock } from "./Message/styled";
+export { TextElement } from "./Messages2.0/styled";
